fix: move bookings endpoint to bookings module
