+++ src/modules/bookings/api/get-bookings.ts
@@ -1,4 +1,4 @@
-import { ENDPOINTS } from '#/modules/home/constants/endpoints.constant.ts';
+import { ENDPOINTS } from '#/modules/bookings/constants/endpoints.constant.ts';
 import { api } from '#/configs';
 import { type Accommodation } from '#/modules/home/types/accommodation.type';
 import { type Booking } from '../types/booking.type';
