+++ src/modules/bookings/constants/endpoints.constant.ts
@@ -1,3 +1,4 @@
 export const ENDPOINTS = {
   cancelBooking: (bookingId: string) => `/bookings/${bookingId}/status`,
+  bookings: '/bookings',
 } as const;
